% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeedlemanWunsch.R
\name{creatingMatrix}
\alias{creatingMatrix}
\title{Creation of the scores matrix}
\usage{
creatingMatrix(matchCost, mismatchCost, gapPenalty, sequenceA, sequenceB)
}
\arguments{
\item{matchCost}{The cost of a match}

\item{mismatchCost}{The cost of a mismatch}

\item{gapPenalty}{The gap penalty(cost of gap/indel)}

\item{sequenceA}{The first nucleotide sequence (positioned on the rows)}

\item{sequenceB}{The second nucleotide sequence (positioned on the columns)}
}
\value{
A matrix containing the scores of the alignment
}
\description{
creatingMatrix creates a matrix containing the scores of the alignment for the NW algorithm
}
\examples{
sequenceA <- "GTT"
sequenceB <- "GCATT"
matchCost <- 7
mismatchCost <- -3
gapPenalty <- -4
scoreMatrix <- creatingMatrix(matchCost,mismatchCost,gapPenalty,sequenceA,sequenceB)

}
