% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeedlemanWunsch.R
\name{globalAlignnmentNeedlemanWunsch}
\alias{globalAlignnmentNeedlemanWunsch}
\title{Needleman Wunsch}
\usage{
globalAlignnmentNeedlemanWunsch(
  matchCost,
  mismatchCost,
  gapPenalty,
  sequenceA,
  sequenceB
)
}
\arguments{
\item{matchCost}{The cost of a match}

\item{mismatchCost}{The cost of a mismatch}

\item{gapPenalty}{The gap penalty(cost of gap/indel)}

\item{sequenceA}{The first nucleotide sequence (positioned on the rows)}

\item{sequenceB}{The second nucleotide sequence (positioned on the columns)}
}
\value{
a list containing the score matrix, the direction matrix, the first alignment, the second alignment
}
\description{
Core function of the package, performs the Needleman-Wunsch algorithm
}
\examples{
sequenceA <- "GTT"
sequenceB <- "GCATT"
matchCost <- 7
mismatchCost <- -3
gapPenalty <- -4
result <- globalAlignnmentNeedlemanWunsch(matchCost,mismatchCost,gapPenalty,sequenceA,sequenceB)

}
