% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeedlemanWunsch.R
\name{performTraceback}
\alias{performTraceback}
\title{Traceback function}
\usage{
performTraceback(matrix, sequenceA, sequenceB)
}
\arguments{
\item{matrix}{The score matrix computed with the two sequences}

\item{sequenceA}{The first nucleotide sequence (positioned on the rows)}

\item{sequenceB}{The second nucleotide sequence (positioned on the columns)}
}
\value{
a list containing the directionMatrix, the first alignment, the second alignment
}
\description{
performTraceback is a function that creates a matrix with the directions of the alignment while computing the alignment itself
}
\examples{
isItAMatch("GTT","GCATT",1,2)

}
